Revert "move Client class to client.py"
This reverts commit 076c7b90005f4215716119737ca6ec4704a8f05e.
